Add deleteTargetPath to deploy

+++ deploy/ssh/deploy.sh
@@ -6,10 +6,11 @@
 # 
 #  Example usage:
 #    #! /bin/sh
-#    host=panacoda.com    # required
-#    port=10022           # optional; default is 22
-#    username=foo         # optional; default is $LOGNAME
-#    targetPath=quux      # optional; default is '', i.e. the home directory
+#    host=panacoda.com     # required
+#    port=10022            # optional; default is 22
+#    username=foo          # optional; default is $LOGNAME
+#    targetPath=quux       # required
+#    deleteTargetPath=true # optional; default is false
 # 
 #    export host port username password targetPath timeout
 # 
@@ -21,8 +22,9 @@ set -euf
 host="${host}"
 port="${port-22}"
 username="${username-$LOGNAME}"
-targetPath="${targetPath-.}"
+targetPath="${targetPath}"
 sourcePath="${1-$sourcePath}"
+deleteTargetPath="${deleteTargetPath-false}"
 
 # TODO implement debugLevel
 
@@ -31,7 +33,15 @@ ssh -o BatchMode=yes -p "$port" "$username@$host" "
   set -euf
   echo # TODO remove this, when we handle debugLevel
 
-  find '$targetPath' -mindepth 1 -maxdepth 1 2>/dev/null | xargs rm -vfR
+  if test -e '$targetPath'; then
+    if test '$deleteTargetPath' = true; then
+      find '$targetPath' -mindepth 1 -maxdepth 1 2>/dev/null | xargs rm -vfR
+    else
+      exec >&2
+      echo 'Error: targetPath already exists, use deleteTargetPath to remove it first'
+      exit 23
+    fi
+  fi
 
   mkdir -pv '$targetPath'
   cd '$targetPath'
